Creation des API pour :  Shift, section, cycle
Creation des API pour :
Shift, section, cycle

+++ backend/src/main/java/org/siges/backend/dao/AcademicPeriodRepository.java
@@ -3,6 +3,6 @@ package org.siges.backend.dao;
 import org.siges.backend.entity.AcademicPeriod;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AcademicPeriodRepository extends JpaRepository<AcademicPeriod, Long> {
+public interface AcademicPeriodRepository extends JpaRepository<AcademicPeriod, Integer> {
 
 }
